wlan:(backport) Fix scan timestamps to android expectations
Marshmallow started to use timestamps in visibility decisions.
If the actual frame timestamp is used, getVisibility in ScanDetailCache
can report a long lost AP as still visible, preventing the switch
to an actually currently available AP.
This fixes the automatic re-connection to an AP, when moving between
different places with different remembered APs.

Change-Id: I1805e4da120c41479f5d8870b2bf4e69baa9f355